finishh detail

+++ SaleDoanInbound/Controllers/ToursController.cs
@@ -2,9 +2,13 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Data.Dtos;
+using Data.Models_QLT;
 using Data.Repository;
 using Microsoft.AspNetCore.Http.Extensions;
 using Microsoft.AspNetCore.Mvc;
+using Microsoft.EntityFrameworkCore.Internal;
+using Newtonsoft.Json;
 using SaleDoanInbound.Models;
 
 namespace SaleDoanInbound.Controllers
@@ -22,7 +26,12 @@ namespace SaleDoanInbound.Controllers
             {
                 Tour = new Data.Models_IB.Tour(),
                 Thanhphos = _unitOfWork.thanhPhoForTuyenTQRepository.GetAll(),
-                Companies = _unitOfWork.khachHangRepository.GetAll()
+                Companies = _unitOfWork.khachHangRepository.GetAll(),
+                Tourkinds = _unitOfWork.tourKindRepository.GetAll(),
+                Dmchinhanhs = _unitOfWork.dmChiNhanhRepository.GetAll(),
+                NguonTours = NguonTour(),
+                listPhong = new List<Data.Models_QLT.Phongban>(),
+                TourDto = new TourDto()
             };
         }
         public IActionResult Index(string searchString = null, int page = 1)
@@ -44,12 +53,12 @@ namespace SaleDoanInbound.Controllers
             //}
 
             var companies = TourVM.Companies;
-            var loaiTours = _unitOfWork.loaiTourRepository.GetAll();
-            var chiNhanhs = _unitOfWork.chiNhanhRepository.GetAll();
+            var loaiTours = TourVM.Tourkinds;
+            var chiNhanhs = TourVM.Dmchinhanhs;
             var cacNoiDungHuyTours = _unitOfWork.cacNoiDungHuyTourRepository.GetAll();
 
             TourVM.TourDtos = _unitOfWork.tourRepository.ListTour(searchString, companies, loaiTours, chiNhanhs, cacNoiDungHuyTours, page);
-            if(TourVM.TourDtos == null)
+            if (TourVM.TourDtos == null)
             {
 
             }
@@ -59,6 +68,11 @@ namespace SaleDoanInbound.Controllers
         public IActionResult Create(string strUrl)
         {
             TourVM.StrUrl = strUrl;
+            TourVM.Tour.ChiNhanhTaoId = 22; // for test
+            ViewBag.tuyenTQ = "BAL,BAN"; //"[BAL,BAN]"; // for test
+            // get list phong ban / thi truong
+            GetListPhongBanDH();
+            // get list phong ban / thi truong
             return View(TourVM);
         }
 
@@ -68,15 +82,63 @@ namespace SaleDoanInbound.Controllers
         {
             if (!ModelState.IsValid)
             {
+                TourVM = new TourViewModel()
+                {
+                    Tour = new Data.Models_IB.Tour(),
+                    Thanhphos = _unitOfWork.thanhPhoForTuyenTQRepository.GetAll(),
+                    Companies = _unitOfWork.khachHangRepository.GetAll(),
+                    Tourkinds = _unitOfWork.tourKindRepository.GetAll(),
+                    Dmchinhanhs = _unitOfWork.dmChiNhanhRepository.GetAll(),
+                    NguonTours = NguonTour()
+                };
                 return View(TourVM);
             }
 
             //TourVM.Tour = new Data.Models_IB.Tour();
+            TourVM.Tour.TuyenTQ = TourVM.Tour.TuyenTQ.Replace(',', '-');
+            // create sgtcode
+
+            var sgtCode = _unitOfWork.tourInfRepository.newSgtcode(Convert.ToDateTime(TourVM.Tour.NgayDen), "STS", TourVM.Tour.PhongBanMaCode);
+            TourVM.Tour.Sgtcode = sgtCode;
+            // create sgtcode
+
             TourVM.Tour.NgayTao = DateTime.Now;
             TourVM.Tour.NguoiTao = "Admin";
+
+            // insert tourinf
+            Tourinf tourinf = new Tourinf();
+
+            tourinf.Sgtcode = TourVM.Tour.Sgtcode;
+            tourinf.Khachle = false;
+            tourinf.CompanyId = TourVM.Tour.MaKH;
+            tourinf.TourkindId = TourVM.Tour.LoaiTourId;
+            tourinf.Arr = TourVM.Tour.NgayDen;
+            tourinf.Dep = TourVM.Tour.NgayDi;
+            tourinf.Pax = TourVM.Tour.SoKhachTT;
+            tourinf.Childern = TourVM.Tour.SKTreEm;
+            tourinf.Reference = TourVM.Tour.TuyenTQ;
+            tourinf.Concernto = "HongVT"; // nguoi tao tour
+            tourinf.Operators = "";
+            tourinf.Departoperator = TourVM.Tour.PhongBanMaCode;
+            tourinf.Departcreate = "IB";
+            tourinf.Routing = "";
+            tourinf.Rate = TourVM.Tour.TyGia;
+            tourinf.Revenue = TourVM.Tour.DoanhThuTT;
+            tourinf.PasstypeId = ""; // tourIB ko co'
+            tourinf.Currency = TourVM.Tour.LoaiTien;
+            tourinf.Chinhanh = "STS"; // chinhanh trien khai
+            tourinf.Chinhanhtao = "STS"; // user login
+            tourinf.Createtour = TourVM.Tour.NgayTao;
+            tourinf.Logfile = TourVM.Tour.LogFile;
+
+            // insert tourinf
+
             try
             {
                 _unitOfWork.tourRepository.Create(TourVM.Tour);
+                // insert tourinf
+                _unitOfWork.tourInfRepository.Create(tourinf);
+                // insert tourinf
                 await _unitOfWork.Complete();
                 SetAlert("Thêm mới thành công.", "success");
                 return Redirect(strUrl);
@@ -89,23 +151,31 @@ namespace SaleDoanInbound.Controllers
 
         }
 
-        public async Task<IActionResult> Edit(int? id, string strUrl)
+        public async Task<IActionResult> Edit(long id, string strUrl)
         {
             TourVM.StrUrl = strUrl;
-            if (id == null)
+            if (id == 0)
                 return NotFound();
 
-            TourVM.Tour = await _unitOfWork.tourRepository.GetByIdAsync(id);
+            TourVM.Tour = await _unitOfWork.tourRepository.GetByLongIdAsync(id);
 
             if (TourVM.Tour == null)
                 return NotFound();
 
+            // gang qua hid tuyentq
+            TourVM.Tour.TuyenTQ = TourVM.Tour.TuyenTQ.Replace('-', ',');
+            // gang qua hid tuyentq
+
+            // get list phong ban / thi truong
+            GetListPhongBanDH();
+            // get list phong ban / thi truong
+
             return View(TourVM);
         }
 
-        [HttpPost]
+        [HttpPost, ActionName("Edit")]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(long id, string strUrl)
+        public async Task<IActionResult> EditPost(long id, string strUrl)
         {
             if (id != TourVM.Tour.Id)
                 return NotFound();
@@ -114,10 +184,44 @@ namespace SaleDoanInbound.Controllers
             {
                 TourVM.Tour.NgaySua = DateTime.Now;
                 TourVM.Tour.NguoiSua = "Admin";
+
+                TourVM.Tour.TuyenTQ = TourVM.Tour.TuyenTQ.Replace(',', '-');
+
+                // update tourinf
+                Tourinf tourinf = new Tourinf();
+
+                tourinf.Sgtcode = TourVM.Tour.Sgtcode;
+                tourinf.Khachle = false;
+                tourinf.CompanyId = TourVM.Tour.MaKH;
+                tourinf.TourkindId = TourVM.Tour.LoaiTourId;
+                tourinf.Arr = TourVM.Tour.NgayDen;
+                tourinf.Dep = TourVM.Tour.NgayDi;
+                tourinf.Pax = TourVM.Tour.SoKhachTT;
+                tourinf.Childern = TourVM.Tour.SKTreEm;
+                tourinf.Reference = TourVM.Tour.TuyenTQ;
+                tourinf.Concernto = "HongVT"; // nguoi tao tour
+                tourinf.Operators = "";
+                tourinf.Departoperator = TourVM.Tour.PhongBanMaCode;
+                tourinf.Departcreate = "IB";
+                tourinf.Routing = "";
+                tourinf.Rate = TourVM.Tour.TyGia;
+                tourinf.Revenue = TourVM.Tour.DoanhThuTT;
+                tourinf.PasstypeId = ""; // tourIB ko co'
+                tourinf.Currency = TourVM.Tour.LoaiTien;
+                tourinf.Chinhanh = "STS"; // chinhanh trien khai
+                tourinf.Chinhanhtao = "STS"; // user login
+                tourinf.Createtour = TourVM.Tour.NgayTao;
+                tourinf.Logfile = TourVM.Tour.LogFile;
+
+                // update tourinf
+
                 try
                 {
 
                     _unitOfWork.tourRepository.Update(TourVM.Tour);
+                    // insert tourinf
+                    _unitOfWork.tourInfRepository.Update(tourinf);
+                    // insert tourinf
                     await _unitOfWork.Complete();
                     SetAlert("Cập nhật thành công", "success");
                     return Redirect(strUrl);
@@ -145,7 +249,62 @@ namespace SaleDoanInbound.Controllers
                 return NotFound();
 
             TourVM.Tour = tour;
+            var tourDto = new TourDto();
 
+            tourDto.Id = tour.Id;
+            tourDto.Sgtcode = tour.Sgtcode;
+            tourDto.KhachLe = tour.KhachLe;
+            tourDto.ChuDeTour = tour.ChuDeTour;
+            tourDto.ThiTruong = tour.ThiTruong;
+            tourDto.NgayKhoa = tour.NgayKhoa;
+            tourDto.NguoiKhoa = tour.NguoiKhoa;
+            tourDto.NgayTao = tour.NgayTao;
+            tourDto.NguoiTao = tour.NguoiTao;
+            tourDto.NgayDen = tour.NgayDen;
+            tourDto.NgayDi = tour.NgayDi;
+            tourDto.TuyenTQ = tour.TuyenTQ;
+            tourDto.SoKhachDK = tour.SoKhachDK;
+            tourDto.DoanhThuDK = tour.DoanhThuDK;
+            //tourDto.CompanyName = companies.Where(x => x.CompanyId == tour.MaKH).FirstOrDefault().Name;
+            tourDto.NgayDamPhan = tour.NgayDamPhan;
+            tourDto.HinhThucGiaoDich = tour.HinhThucGiaoDich;
+            tourDto.NgayKyHopDong = tour.NgayKyHopDong;
+            tourDto.NguoiKyHopDong = tour.NguoiKyHopDong;
+            tourDto.HanXuatVe = tour.HanXuatVe;
+            tourDto.NgayThanhLyHD = tour.NgayThanhLyHD;
+            tourDto.SoKhachTT = tour.SoKhachTT;
+            tourDto.SKTreEm = tour.SKTreEm;
+            tourDto.DoanhThuTT = tour.DoanhThuTT;
+            tourDto.ChuongTrinhTour = tour.ChuongTrinhTour;
+            tourDto.NoiDungThanhLyHD = tour.NoiDungThanhLyHD;
+            tourDto.DichVu = tour.DichVu;
+            tourDto.DaiLy = tour.DaiLy;
+            tourDto.TrangThai = tour.TrangThai;
+            tourDto.NgaySua = tour.NgaySua;
+            tourDto.NguoiSua = tour.NguoiSua;
+            tourDto.TenLoaiTour = (tour.LoaiTourId == 0) ? "" : _unitOfWork.tourKindRepository.GetById(tour.LoaiTourId.Value).TourkindInf;
+            tourDto.MaCNTao = (tour.ChiNhanhTaoId == 0) ? "" : TourVM.Dmchinhanhs.Where(x => x.Id == tour.ChiNhanhTaoId).FirstOrDefault().Macn;
+            tourDto.NgayNhanDuTien = tour.NgayNhanDuTien;
+            tourDto.LyDoNhanDu = tour.LyDoNhanDu;
+            tourDto.SoHopDong = tour.SoHopDong;
+            tourDto.LaiChuaVe = tour.LaiChuaVe;
+            tourDto.LaiGomVe = tour.LaiGomVe;
+            tourDto.LaiThucTeGomVe = tour.LaiThucTeGomVe;
+            tourDto.NguonTour = tour.NguonTour;
+            tourDto.FileKhachDiTour = tour.FileKhachDiTour;
+            tourDto.FileVeMayBay = tour.FileVeMayBay;
+            tourDto.FileBienNhan = tour.FileBienNhan;
+            tourDto.NguoiDaiDien = tour.NguoiDaiDien;
+            tourDto.DoiTacNuocNgoai = tour.DoiTacNuocNgoai;
+            tourDto.MaCNDH = TourVM.Dmchinhanhs.Where(x => x.Id == tour.ChiNhanhDHId).FirstOrDefault().Macn;
+            tourDto.NgayHuyTour = tour.NgayHuyTour;
+            tourDto.NDHuyTour = (tour.NDHuyTourId == 0) ? "" : _unitOfWork.cacNoiDungHuyTourRepository.GetById(tour.NDHuyTourId).NoiDung;
+            tourDto.GhiChu = tour.GhiChu;
+            tourDto.LoaiTien = tour.LoaiTien;
+            tourDto.TyGia = tour.TyGia;
+            tourDto.LogFile = tour.LogFile;
+
+            TourVM.TourDto = tourDto;
             return View(TourVM);
         }
 
@@ -156,8 +315,16 @@ namespace SaleDoanInbound.Controllers
             var tour = _unitOfWork.tourRepository.GetById(id);
             if (tour == null)
                 return NotFound();
+
+            // tourinf
+            var tourInf = await _unitOfWork.tourInfRepository.GetByIdAsync(tour.Sgtcode);
+            if (tourInf == null)
+                return NotFound();
+            // tourinf
+
             try
             {
+                _unitOfWork.tourInfRepository.Delete(tourInf);
                 _unitOfWork.tourRepository.Delete(tour);
                 await _unitOfWork.Complete();
                 SetAlert("Xóa thành công.", "success");
@@ -183,10 +350,65 @@ namespace SaleDoanInbound.Controllers
         //    }
         //}
 
-        private List<string> Visa()
+        private List<ListViewModel> Visa()
         {
-            return new List<string>()
-            { "Khong visa", "Nuoc ngoai", "Cua khau" };
+            return new List<ListViewModel>()
+            {
+                new ListViewModel(){id = 1, Name = "Không có"},
+                new ListViewModel(){id = 2, Name = "Nước ngoài"},
+                new ListViewModel(){id = 3, Name = "Cửa khẩu"},
+            };
+
+        }
+
+        private List<ListViewModel> NguonTour()
+        {
+            return new List<ListViewModel>()
+            {
+                new ListViewModel(){id = 1, Name = "Nội bộ" },
+                new ListViewModel(){id = 2, Name = "TMDT" },
+
+            };
+        }
+
+        public async Task<JsonResult> GetKHByMaKH(string maKH)
+        {
+            var khachHang = JsonConvert.SerializeObject(await _unitOfWork.khachHangRepository.GetByIdAsync(maKH));
+            return Json(new
+            {
+                khachHang = khachHang
+            });
+        }
+
+        private void GetListPhongBanDH()
+        {
+
+            // get list maphong after distinct space and split ','
+            var phongbans = _unitOfWork.phongBanRepository.GetAll();
+            var maPhongs = phongbans.Select(x => x.Macode).Distinct();
+
+            var listString = new List<string>();
+            foreach (var item in maPhongs)
+            {
+                var itemArray = item.Split(',');
+                for (int i = 0; i < itemArray.Length; i++)
+                {
+                    listString.Add(itemArray[i]);
+
+                }
+            }
+
+            int j = 1;
+            foreach (var maCode in listString)
+            {
+                TourVM.listPhong.Add(new Phongban()
+                {
+                    Maphong = phongbans.Where(x => x.Macode.Contains(maCode)).FirstOrDefault().Maphong,
+                    Macode = maCode
+                });
+
+            }
+            // get list maphong after distinct space and split ','
         }
     }
 }
